Refactor: reducer to slice
